fix -Wsign-compare warnings in modules

+++ display/d.linegraph/main.c
@@ -386,7 +386,7 @@ int main(int argc, char **argv)
 
     /* TODO: put this to variables, and avoid -Wsign-compare */
     for (i = 0; i < 3; i++) {
-        for (j = 0; j < strlen(title[i]->answer); j++)
+        for (j = 0; (size_t)j < strlen(title[i]->answer); j++)
             if (title[i]->answer[j] == '_')
                 title[i]->answer[j] = ' ';
     }
@@ -519,9 +519,9 @@ int main(int argc, char **argv)
     c = 0;
     j = 1;
     if (y_color_opt->answer != NULL) {
-        for (i = 0; i <= (strlen(y_color_opt->answer)); i++) {
+        for (i = 0; (size_t)i <= (strlen(y_color_opt->answer)); i++) {
             if ((y_color_opt->answer[i] == ',') ||
-                (i == (strlen(y_color_opt->answer)))) {
+                ((size_t)i == (strlen(y_color_opt->answer)))) {
                 color_name[c] = '\0';
                 in[j].color = D_translate_color(color_name);
                 j++;
